Improve consistency in policy checks
Also simplifies policy defaults.

+++ views/post/partials/actions.blade.php
@@ -8,8 +8,10 @@
             <div class="mb-3">
                 <label for="action">{{ trans_choice('forum::general.actions', 1) }}</label>
                 <select name="action" id="action" class="form-control">
-                    @can ('delete', $post)
-                        <option value="delete" data-confirm="true" data-method="delete">{{ trans('forum::general.delete') }}</option>
+                    @can ('deletePosts', $post->thread)
+                        @can ('delete', $post)
+                            <option value="delete" data-confirm="true" data-method="delete">{{ trans('forum::general.delete') }}</option>
+                        @endcan
                     @endcan
                 </select>
             </div>
